+++ tests/DependencyInjection/CQRSExtensionTest.php
@@ -34,7 +34,7 @@ final class CQRSExtensionTest extends TestCase
         // No data is supplied as config, therefore the parameters are set, but empty
         self::assertNull($container->getParameter('cqrs.query_controller.default_request_decoder_class'));
         /** @psalm-suppress PossiblyInvalidArgument */
-        self::assertCount(0, $container->getParameter('cqrs.query_controller.default_dto_data_transformer_classes'));
+        self::assertCount(0, $container->getParameter('cqrs.query_controller.default_request_data_transformer_classes'));
         self::assertNull($container->getParameter('cqrs.query_controller.default_dto_constructor_class'));
         /** @psalm-suppress PossiblyInvalidArgument */
         self::assertCount(0, $container->getParameter('cqrs.query_controller.default_dto_validator_classes'));
@@ -44,7 +44,7 @@ final class CQRSExtensionTest extends TestCase
 
         self::assertNull($container->getParameter('cqrs.command_controller.default_request_decoder_class'));
         /** @psalm-suppress PossiblyInvalidArgument */
-        self::assertCount(0, $container->getParameter('cqrs.command_controller.default_dto_data_transformer_classes'));
+        self::assertCount(0, $container->getParameter('cqrs.command_controller.default_request_data_transformer_classes'));
         self::assertNull($container->getParameter('cqrs.command_controller.default_dto_constructor_class'));
         /** @psalm-suppress PossiblyInvalidArgument */
         self::assertCount(0, $container->getParameter('cqrs.command_controller.default_dto_validator_classes'));