+++ tests/RequestDecoder/JsonRequestDecoderTest.php
@@ -27,16 +27,16 @@ final class JsonRequestDecoderTest extends TestCase
             content: $json,
         );
 
-        $expectedDTOData = [
+        $expectedRequestData = [
             'userId' => 'abf6b545-951e-46d8-b444-dc57b31ee51f',
             'amount' => 2000,
             'isRelevant' => true,
         ];
 
         // -- Act
-        $dtoData = $jsonRequestDecoder->decodeRequest($request);
+        $requestData = $jsonRequestDecoder->decodeRequest($request);
 
         // -- Assert
-        self::assertSame($expectedDTOData, $dtoData);
+        self::assertSame($expectedRequestData, $requestData);
     }
 }